ModelWrappers

Documentation for ModelWrappers.

ModelWrappers.max_val

ModelWrappers.min_Δ

ModelWrappers.AbstractConstraint

ModelWrappers.AbstractInitialization

ModelWrappers.BaseModel

ModelWrappers.Constrained

ModelWrappers.Fixed

ModelWrappers.FlattenConstructor

ModelWrappers.FlattenDefault

ModelWrappers.FlattenTypes

ModelWrappers.ModelName

ModelWrappers.ModelWrapper

ModelWrappers.NoInitialization

ModelWrappers.Objective

ModelWrappers.Objective

ModelWrappers.Objective

ModelWrappers.OptimInitialization

ModelWrappers.Param

ModelWrappers.ParameterInfo

ModelWrappers.PriorInitialization

ModelWrappers.ReConstructor

ModelWrappers.Tagged

ModelWrappers.TransformConstructor

ModelWrappers.Unconstrained

ModelWrappers.UnflattenConstructor

ModelWrappers.UnflattenTypes

Base.fill

Base.fill!

Base.print

BaytesCore.generate

BaytesCore.generate_showvalues

ModelWrappers.Simplex_from_flatten

ModelWrappers.Simplex_from_flatten!

ModelWrappers.Symmetric_from_flatten

ModelWrappers.Symmetric_from_flatten!

ModelWrappers._allparam

ModelWrappers._anyparam

ModelWrappers._check

ModelWrappers._checkfinite

ModelWrappers._checkparams

ModelWrappers._checkprior

ModelWrappers._checksampleable

ModelWrappers._get_constraint

ModelWrappers._get_val

ModelWrappers._paramnames

ModelWrappers.check_constraint

ModelWrappers.constrain

ModelWrappers.constrain

ModelWrappers.constrain

ModelWrappers.constrain!

ModelWrappers.construct_flatten

ModelWrappers.dynamics

ModelWrappers.fill_array!

ModelWrappers.flatten

ModelWrappers.flatten

ModelWrappers.flattenAD

ModelWrappers.flatten_Simplex

ModelWrappers.flatten_Symmetric

ModelWrappers.flatten_array

ModelWrappers.log_abs_det_jac

ModelWrappers.log_abs_det_jac

ModelWrappers.log_prior

ModelWrappers.log_prior

ModelWrappers.log_prior_with_transform

ModelWrappers.log_prior_with_transform

ModelWrappers.paramcount

ModelWrappers.paramnames

ModelWrappers.predict

ModelWrappers.predictive

ModelWrappers.sample_constraint

ModelWrappers.simulate

ModelWrappers.tag

ModelWrappers.unconstrain

ModelWrappers.unconstrain

ModelWrappers.unconstrain!

ModelWrappers.unconstrain_flatten

ModelWrappers.unflatten

ModelWrappers.unflatten

ModelWrappers.unflatten!

ModelWrappers.unflattenAD

ModelWrappers.unflatten_constrain

ModelWrappers.unflatten_constrain!

StatsBase.sample

StatsBase.sample!

ModelWrappers.max_val — Constant

Maximum value after transformation before tagged as non-finite.

ModelWrappers.min_Δ — Constant

Smallest decrease allowed in the log objective results before tagged as divergent.

ModelWrappers.AbstractConstraint — Type

abstract type AbstractConstraint

Abstract super type for parameter constraints.

ModelWrappers.AbstractInitialization — Type

abstract type AbstractInitialization

Abstract method to initialize parameter for individual kernels.

Fields

ModelWrappers.BaseModel — Type

struct BaseModel <: ModelName

Default modelname of Baytes.Model struct.

ModelWrappers.Constrained — Type

struct Constrained{B<:Bijection} <: AbstractConstraint

Utility struct to help assign boundaries to parameter - keeps scalar parameter constrained.

Fields

bijection::Bijection

ModelWrappers.Fixed — Type

struct Fixed{B<:Bijection} <: AbstractConstraint

Utility struct to help assign boundaries to parameter - keeps parameter fixed. Useful for assigning buffer values for functions of parameter.

Fields

bijection::Bijection

ModelWrappers.FlattenConstructor — Type

struct FlattenConstructor{S<:Function, T<:Function}

Contains information for flatten construct.

Fields

strict::Function

flexible::Function

ModelWrappers.FlattenDefault — Type

struct FlattenDefault{T<:AbstractFloat, F<:FlattenTypes}

Default arguments for flatten function.

Fields

output::Type{T} where T<:AbstractFloat

Type of flatten output

flattentype::FlattenTypes

Determines if all inputs are flattened (FlattenAll) or only continuous values (FlattenContinuous).

ModelWrappers.FlattenTypes — Type

abstract type FlattenTypes

Supertype for dispatching different types of flatten. Determines if all inputs are flattened (FlattenAll) or only continuous values (FlattenContinuous).

Fields

ModelWrappers.ModelName — Type

abstract type ModelName

Abstract super type for Baytes Models.

ModelWrappers.ModelWrapper — Type

mutable struct ModelWrapper{M<:(Union{ModelName, P} where P), A<:NamedTuple, C<:NamedTuple, B<:ParameterInfo} <: BaytesCore.AbstractModelWrapper

Baytes Model struct.

Contains information about current Model value, name, and information, see also ParameterInfo.

Fields

val::NamedTuple

Current Model values as NamedTuple - works with Nested Tuples.

arg::NamedTuple

Supplementary arguments for log target function that are fixed and dont need to be stored in a trace.

info::ParameterInfo

Information about parameter distributions, transformations and constraints, see ParameterInfo.

id::Union{ModelName, P} where P

Model id, per default BaseModel. Useful for dispatching ModelWrapper struct.

ModelWrappers.NoInitialization — Type

Use current model.val parameter as initial parameter

ModelWrappers.Objective — Type

struct Objective{M<:ModelWrapper, D, T<:Tagged, F<:AbstractFloat} <: BaytesCore.AbstractObjective

Functor to calculate 'ℓfunc' and gradient at unconstrained 'θᵤ', including eventual Jacobian adjustments.

Fields

model::ModelWrapper

data::Any

tagged::Tagged

temperature::AbstractFloat

ModelWrappers.Objective — Method

Functor to call target function for Model given parameter and data. Default method to be used in Automatic Differentiation. model.arg and data are arguments so they can be declared as constant with Enzyme AD engine.

Examples

ModelWrappers.Objective — Method

Functor to call target function for Model given parameter and data.

Examples

ModelWrappers.OptimInitialization — Type

Use custom optimization technique for initialization.

ModelWrappers.Param — Type

struct Param{A, B}

Utility struct to define Parameter in a way ModelWrappers.jl can handle. Will be separated in ModelWrapper struct for type stability.

Fields

constraint::Any

val::Any

ModelWrappers.ParameterInfo — Type

struct ParameterInfo{R<:ReConstructor, U<:ReConstructor, T<:TransformConstructor}

Contains information about parameter distributions, transformations and constraints.

Fields

reconstruct::ReConstructor

Contains information for flatten/unflatten parameter

reconstructᵤ::ReConstructor

Contains information to reconstruct unconstrained parameter - important for non-bijective transformations

transform::TransformConstructor

Contains information for constraining and unconstraining parameter.

ModelWrappers.PriorInitialization — Type

Sample (up to Ntrials) times from prior and check if log target distribution is finite at proposed parameter in unconstrained space.

ModelWrappers.ReConstructor — Type

struct ReConstructor{F<:FlattenDefault, S<:FlattenConstructor, T<:UnflattenConstructor}

Contains information for flatten/unflatten construct.

Fields

default::FlattenDefault

flatten::FlattenConstructor

unflatten::UnflattenConstructor

ModelWrappers.Tagged — Type

struct Tagged{A<:NamedTuple, B<:ParameterInfo}

Stores information for a subset of 'model' parameter.

Fields

parameter::NamedTuple

Subset of ModelWrapper parameter names.

info::ParameterInfo

Information about subset of parameter distributions, transformations and constraints, see ParameterInfo.

ModelWrappers.TransformConstructor — Type

TransformConstructor(x )

Contains information to constrain and unconstrain parameter for all parameter.

Examples

ModelWrappers.Unconstrained — Type

struct Unconstrained{B<:Bijection} <: AbstractConstraint

Utility struct to help assign boundaries to parameter - keeps parameter unconstrained. Useful for assigning buffer values for functions of parameter.

Fields

bijection::Bijection

ModelWrappers.UnflattenConstructor — Type

struct UnflattenConstructor{S<:Function, T<:Function}

Contains information for unflatten construct.

Fields

strict::Function

flexible::Function

ModelWrappers.UnflattenTypes — Type

abstract type UnflattenTypes

Determines if unflatten returns original type or if type may change (AD friendly).

Fields

Base.fill! — Method

fill!(model, θ)

Inplace version of fill.

Examples

Base.fill — Method

fill(model, θ)

Fill 'model' values with NamedTuple 'θ'.

Examples

Base.print — Method

print(model)
print(model, params)

Print 'model' parameter values and constraints of symbols 'params'.

Examples

BaytesCore.generate — Method

generate(_rng, objective)

Generate statistics given model parameter and data.

Examples

BaytesCore.generate_showvalues — Method

generate_showvalues(model)

Show current values of Model as NamedTuple

Examples

ModelWrappers.Simplex_from_flatten! — Method

Simplex_from_flatten!(buffer, x_vec)

Inplace version of Simplexfromflatten. Not exported.

Examples

ModelWrappers.Simplex_from_flatten — Method

Expand vector of k-1 dimensions back to k dimensions. Not exported.

Examples

ModelWrappers.Symmetric_from_flatten! — Method

Symmetric_from_flatten!(mat, x_vec, idx)

Inplace version of Symmetricfromflatten. Not exported.

Examples

ModelWrappers.Symmetric_from_flatten — Method

Symmetric_from_flatten(x_vec, idx)

Expand vector back to (symmetric) Matrix. Not exported.

Examples

ModelWrappers._allparam — Method

_allparam(val)

Returns NamedTuple of true/false given parameter is not fixed. Not exported.

Examples

ModelWrappers._anyparam — Method

_anyparam(val)

Returns NamedTuple of true/false given parameter is not fixed. Not exported.

Examples

ModelWrappers._check — Method

Check if (constraint, val) combination is valid. If nothing else specified, returns false (!) per default so check necessary.

ModelWrappers._checkfinite — Method

_checkfinite(θ)
_checkfinite(θ, max_val)

Check if 'θ' is of finite value and return Bool. Not exported.

Examples

ModelWrappers._checkparams — Method

_checkparams(param)

Check if all values in (Nested) NamedTuple are a 'Param' struct and return Bool. Not exported.

Examples

ModelWrappers._checkprior — Method

_checkprior(prior)

Check if 'prior' is a valid density and return Bool. Not exported.

Examples

ModelWrappers._checksampleable — Method

_checksampleable(constraint)

Check if argument is not fixed. Returns NamedTuple with true/false. Needed in addition to _checkprior for nested NamedTuples. Not exported.

Examples

ModelWrappers._get_constraint — Method

_get_constraint(param)

Recursively collect constraints of 'Param' struct. Not exported.

Examples

ModelWrappers._get_val — Method

_get_val(param)

Recursively collect values of 'Param' struct. Not exported.

Examples

ModelWrappers._paramnames — Method

_paramnames(sym, len)

Return parameter names as a string in increasing order. Not exported.

Examples

ModelWrappers.check_constraint — Method

check_constraint(x )

Check if constrain and unconstrain functions of 'constraint' can map 'val' correctly. Will be called when initiating a 'Param' struct.

Examples

ModelWrappers.constrain — Function

Constrain val with given constraint

ModelWrappers.constrain! — Method

Inplace constrain val with given constraint, using 'val' as buffer.

ModelWrappers.constrain — Method

constrain(model, θ)

Constrain 'θᵤ' values with model.info ParameterInfo.

Examples

ModelWrappers.constrain — Method

constrain(model, tagged, θ)

Constrain 'θᵤ' values with tagged.info ParameterInfo.

Examples

ModelWrappers.construct_flatten — Function

construct_flatten(x ) Construct a flatten function for 'x' given specifications in 'df'.

Examples

ModelWrappers.dynamics — Method

dynamics(objective)

Assign model dynamics for a given objective.

Examples

ModelWrappers.fill_array! — Method

fill_array!(buffer, vec)

Fill array with elements of vec. Not exported.

Examples

ModelWrappers.flatten — Function

flatten(x )

Convert 'x' into a Vector.

Examples

ModelWrappers.flatten — Method

flatten(model)

Flatten 'model' values and return as vector.

Examples

ModelWrappers.flattenAD — Function

flattenAD(x )

Convert 'x' into a Vector that is AD compatible.

Examples

ModelWrappers.flatten_Simplex — Method

flatten_Simplex(x)

Flatten vector x to k-1 dimensions. Not exported.

Examples

ModelWrappers.flatten_Symmetric — Method

flatten_Symmetric(mat, idx)

Flatten matrix to vector. Not exported.

Examples

ModelWrappers.flatten_array — Method

flatten_array(mat)

Flatten array x. Not exported.

Examples

ModelWrappers.log_abs_det_jac — Function

Compute log(abs(determinant(jacobian(x)))) for given transformer to unconstrained (!) domain.

ModelWrappers.log_abs_det_jac — Method

log_abs_det_jac(model)

Evaluate eventual Jacobian adjustments from transformations at 'model' values.

Examples

ModelWrappers.log_prior — Method

log_prior(model)

Evaluate Log density of 'model' prior given current 'model' values.

Examples

ModelWrappers.log_prior — Method

log_prior(prior, θ)

Evaluate Log density of 'prior' at 'θ'.

Examples

ModelWrappers.log_prior_with_transform — Method

log_prior_with_transform(model)

Evaluate Log density and eventual Jacobian adjustments of 'model' prior given current 'model' values.

Examples

ModelWrappers.log_prior_with_transform — Method

Evaluate Log density and eventual Jacobian adjustments from transformation of 'prior' at 'θ'.

Examples

ModelWrappers.paramcount — Method

paramcount(sym, types, val)

Count length of nested parameter tuple. Not exported.

Examples

ModelWrappers.paramnames — Method

paramnames(sym, types, constraint, val)

Return all parameter names in increasing order. Not exported.

Examples

ModelWrappers.predict — Method

predict(_rng, objective)

Predict new data given model parameter and data.

Examples

ModelWrappers.predictive — Method

predictive(_rng, objective, init, iter)

Use Prior predictive samples to check model assumptions. Needs dispatch on simulate(rng, model).

Examples

ModelWrappers.sample_constraint — Method

sample_constraint(_rng, prior, val)

Sample from constraint if 'prior'.

Examples

ModelWrappers.simulate — Function

Simulate data given Model parameter.

Examples

ModelWrappers.tag — Method

tag(mat)
tag(mat, upper)
tag(mat, upper, diag)

Assign subset of elements to track in Matrix mat. Not exported.

Examples

ModelWrappers.unconstrain — Function

Unconstrain val with given constraint

ModelWrappers.unconstrain! — Method

Inplace unconstrain val with given constraint, using 'val' as buffer.

ModelWrappers.unconstrain — Method

unconstrain(model)

Unconstrain 'model' values and return as NamedTuple.

Examples

ModelWrappers.unconstrain_flatten — Method

unconstrain_flatten(model)

Flatten and unconstrain 'model' values and return as vector.

Examples

ModelWrappers.unflatten — Function

unflatten(x )

Unflatten 'x' into original shape.

Examples

ModelWrappers.unflatten! — Method

unflatten!(model, θ)

Inplace version of unflatten.

Examples

ModelWrappers.unflatten — Method

Unlatten Vector 'θ' given constraints from 'model' and return as NamedTuple.

Examples

ModelWrappers.unflattenAD — Function

unflattenAD(x )

Unflatten 'x' into original shape but keep type information of 'x' for AD compatibility.

Examples

ModelWrappers.unflatten_constrain! — Method

Inplace version of unflatten_constrain.

Examples

ModelWrappers.unflatten_constrain — Method

Constrain and Unflatten vector 'θᵤ' given 'model' constraints.

Examples

StatsBase.sample! — Method

sample!(_rng, model)

Inplace version of sample.

Examples

StatsBase.sample — Method

sample(_rng, model)

Sample from 'model' prior and return as NamedTuple.

Examples