ModelWrappers

Documentation for ModelWrappers.

ModelWrappers.max_val

ModelWrappers.min_Δ

ModelWrappers.AbstractConstraint

ModelWrappers.AbstractDifferentiableTune

ModelWrappers.AnalyticalDiffTune

ModelWrappers.AutomaticDifferentiationMethod

ModelWrappers.BaseModel

ModelWrappers.Constrained

ModelWrappers.DiffObjective

ModelWrappers.Fixed

ModelWrappers.FlattenDefault

ModelWrappers.FlattenTypes

ModelWrappers.ModelName

ModelWrappers.ModelWrapper

ModelWrappers.Objective

ModelWrappers.Objective

ModelWrappers.Param

ModelWrappers.ParameterInfo

ModelWrappers.Tagged

ModelWrappers.Unconstrained

ModelWrappers.UnflattenTypes

ModelWrappers.ℓDensityResult

ModelWrappers.ℓGradientResult

ModelWrappers.ℓObjectiveResult

Base.fill

Base.fill!

ModelWrappers.Simplex_from_flatten

ModelWrappers.Simplex_from_flatten!

ModelWrappers.Symmetric_from_flatten

ModelWrappers.Symmetric_from_flatten!

ModelWrappers._allparam

ModelWrappers._anyparam

ModelWrappers._checkfinite

ModelWrappers._checkparam

ModelWrappers._checkparams

ModelWrappers._checkprior

ModelWrappers._checksampleable

ModelWrappers._config

ModelWrappers._diffresults_buffer

ModelWrappers._get_constraint

ModelWrappers._get_val

ModelWrappers._paramnames

ModelWrappers._to_bijector

ModelWrappers._to_inv_bijector

ModelWrappers.check_gradients

ModelWrappers.constrain

ModelWrappers.dynamics

ModelWrappers.fill_array!

ModelWrappers.flatten

ModelWrappers.flatten

ModelWrappers.flatten_Simplex

ModelWrappers.flatten_Symmetric

ModelWrappers.flatten_array

ModelWrappers.generate

ModelWrappers.log_abs_det_jac

ModelWrappers.log_abs_det_jac

ModelWrappers.log_density

ModelWrappers.log_density_and_gradient

ModelWrappers.log_density_and_hessian

ModelWrappers.log_prior

ModelWrappers.log_prior

ModelWrappers.log_prior_with_transform

ModelWrappers.log_prior_with_transform

ModelWrappers.paramcount

ModelWrappers.paramnames

ModelWrappers.predict

ModelWrappers.sample_constraint

ModelWrappers.simulate

ModelWrappers.tag

ModelWrappers.unconstrain

ModelWrappers.unconstrain

ModelWrappers.unconstrain_flatten

ModelWrappers.unflatten

ModelWrappers.unflatten!

ModelWrappers.unflatten_constrain

ModelWrappers.unflatten_constrain!

StatsBase.sample

StatsBase.sample!

ModelWrappers.max_val — Constant

Maximum value after transformation before tagged as non-finite.

ModelWrappers.min_Δ — Constant

Smallest decrease allowed in the log objective results before tagged as divergent.

ModelWrappers.AbstractConstraint — Type

abstract type AbstractConstraint

Abstract super type for parameter constraints.

ModelWrappers.AbstractDifferentiableTune — Type

abstract type AbstractDifferentiableTune

Abstract super type for Tuning structs of differentiable functions.

ModelWrappers.AnalyticalDiffTune — Type

struct AnalyticalDiffTune{G<:Function} <: ModelWrappers.AbstractDifferentiableTune

Stores information for evaluating and taking the gradient of an objective function.

Fields

gradient::Function

Gradient as function of ℓobjective and parameter vector in unconstrained space, gradient(ℓobjective, θᵤ).

ModelWrappers.AutomaticDifferentiationMethod — Type

abstract type AutomaticDifferentiationMethod

Abstract super type for Supported Automatic Differentiation backends.

ModelWrappers.BaseModel — Type

struct BaseModel <: ModelName

Default modelname of Baytes.Model struct.

ModelWrappers.Constrained — Type

struct Constrained{T<:Real} <: AbstractConstraint

Utility struct to help assign boundaries to parameter - keeps scalar parameter constrained.

Fields

min::Real

max::Real

ModelWrappers.DiffObjective — Type

struct DiffObjective{O<:Objective, T<:ModelWrappers.AbstractDifferentiableTune}

Objective struct with additional information about AD backend and configuration.

Fields

objective::Objective

Objective as function of a parameter vector in unconstrained space.

tune::ModelWrappers.AbstractDifferentiableTune

Automatic Differentiation configurations.

ModelWrappers.Fixed — Type

struct Fixed <: AbstractConstraint

Utility struct to help assign boundaries to parameter - keeps parameter fixed. Useful for assigning buffer values for functions of parameter.

Fields

ModelWrappers.FlattenDefault — Type

struct FlattenDefault{T<:AbstractFloat, F<:FlattenTypes, S<:UnflattenTypes}

Default arguments for flatten function.

Fields

output::Type{T} where T<:AbstractFloat

Type of flatten output

flattentype::FlattenTypes

Determines if all inputs are flattened (FlattenAll) or only continuous values (FlattenContinuous).

unflattentype::UnflattenTypes

Determines if unflatten returns original type or if type may change (AD friendly).

ModelWrappers.FlattenTypes — Type

Supertype for dispatching different types of flatten. Determines if all inputs are flattened (FlattenAll) or only continuous values (FlattenContinuous).

ModelWrappers.ModelName — Type

abstract type ModelName

Abstract super type for Baytes Models.

ModelWrappers.ModelWrapper — Type

mutable struct ModelWrapper{M<:(Union{ModelName, P} where P), A<:NamedTuple, B<:ParameterInfo} <: BaytesCore.AbstractModelWrapper

Baytes Model struct.

Contains information about current Model value, name, and information, see also ParameterInfo.

Fields

val::NamedTuple

Current Model values as NamedTuple - works with Nested Tuples.

info::ParameterInfo

Information about parameter distributions, transformations and constraints, see ParameterInfo.

id::Union{ModelName, P} where P

Model id, per default BaseModel. Useful for dispatching ModelWrapper struct.

ModelWrappers.Objective — Type

struct Objective{M<:ModelWrapper, D, T<:Tagged, F<:AbstractFloat} <: BaytesCore.AbstractObjective

Functor to calculate 'ℓfunc' and gradient at unconstrained 'θᵤ', including eventual Jacobian adjustments.

Fields

model::ModelWrapper

data::Any

tagged::Tagged

temperature::AbstractFloat

ModelWrappers.Objective — Method

Functor to call target function for Model given parameter and data.

Examples

ModelWrappers.Param — Type

struct Param{A, B}

Utility struct to define Parameter in a way ModelWrappers.jl can handle. Will be separated in ModelWrapper struct for type stability.

Fields

val::Any

constraint::Any

ModelWrappers.ParameterInfo — Type

struct ParameterInfo{A, B, C, D<:FlattenDefault, E<:Function, F<:Function}

Contains information about parameter distributions, transformations and constraints.

Fields

constraint::Any

Constraint distribution/boundaries for all model parameter.

b::Any

Bijector for all model parameter.

b⁻¹::Any

Inverse-Bijector for all model parameter.

flattendefault::FlattenDefault

Default Flattening setting

unflatten::Function

Function to unflatten model parameter, if provided as a vector.

unflatten_AD::Function

Function to unflatten model parameter, if provided as a vector.

ModelWrappers.Tagged — Type

struct Tagged{A<:NamedTuple, B<:ParameterInfo}

Stores information for a subset of 'model' parameter.

Fields

parameter::NamedTuple

Subset of ModelWrapper parameter names.

info::ParameterInfo

Information about subset of parameter distributions, transformations and constraints, see ParameterInfo.

ModelWrappers.Unconstrained — Type

struct Unconstrained <: AbstractConstraint

Utility struct to help assign boundaries to parameter - keeps parameter unconstrained. Useful for assigning buffer values for functions of parameter.

Fields

ModelWrappers.UnflattenTypes — Type

Determines if unflatten returns original type or if type may change (AD friendly).

ModelWrappers.ℓDensityResult — Type

struct ℓDensityResult{T, S} <: ℓObjectiveResult

Stores result for log density and parameter for 'ℓobjective' evaluation at 'parameter'.

Fields

θᵤ::Any

Parameter in unconstrained space.

ℓθᵤ::Any

Log density at θᵤ.

ModelWrappers.ℓGradientResult — Type

struct ℓGradientResult{T, S, G} <: ℓObjectiveResult

Stores result for log density, gradient, and parameter for 'ℓobjective' evaluation at 'parameter'.

Fields

θᵤ::Any

Parameter in unconstrained space.

ℓθᵤ::Any

Log density at θᵤ.

∇ℓθᵤ::Any

Gradient of log density at θᵤ.

ModelWrappers.ℓObjectiveResult — Type

abstract type ℓObjectiveResult <: BaytesCore.AbstractResult

Abstract super type for AbstractDifferentiableObjective results.

Base.fill! — Method

fill!(model, θ)

Inplace version of fill.

Examples

Base.fill — Method

fill(model, θ)

Fill 'model' values with NamedTuple 'θ'.

Examples

ModelWrappers.Simplex_from_flatten! — Method

Simplex_from_flatten!(buffer, x_vec)

Inplace version of Simplexfromflatten. Not exported.

Examples

ModelWrappers.Simplex_from_flatten — Method

Simplex_from_flatten(x_vec)

Expand vector of k-1 dimensions back to k dimensions. Not exported.

Examples

ModelWrappers.Symmetric_from_flatten! — Method

Symmetric_from_flatten!(mat, x_vec, idx)

Inplace version of Symmetricfromflatten. Not exported.

Examples

ModelWrappers.Symmetric_from_flatten — Method

Symmetric_from_flatten(x_vec, idx)

Expand vector back to (symmetric) Matrix. Not exported.

Examples

ModelWrappers._allparam — Method

_allparam(val)

Returns NamedTuple of true/false given parameter is not fixed. Not exported.

Examples

ModelWrappers._anyparam — Method

_anyparam(val)

Returns NamedTuple of true/false given parameter is not fixed. Not exported.

Examples

ModelWrappers._checkfinite — Method

_checkfinite(θ)
_checkfinite(θ, max_val)

Check if 'θ' is of finite value and return Bool. Not exported.

Examples

ModelWrappers._checkparam — Method

_checkparam(_rng, val, constraint)

Check if provided val-constraint combination is valid for Param struct. Not exported.

Examples

ModelWrappers._checkparams — Method

_checkparams(param)

Check if all values in (Nested) NamedTuple are a 'Param' struct and return Bool. Not exported.

Examples

ModelWrappers._checkprior — Method

_checkprior(prior)

Check if 'prior' is a valid density and return Bool. Not exported.

Examples

ModelWrappers._checksampleable — Method

_checksampleable(constraint)

Check if argument is not fixed. Returns NamedTuple with true/false. Needed in addition to _checkprior for nested NamedTuples. Not exported.

Examples

ModelWrappers._config — Method

_config(differentiation, objective, θᵤ)

Write config file for AD wrapper.

Examples

ModelWrappers._diffresults_buffer — Method

Initiate DiffResults.MutableDiffResult struct. Not exported.

Examples

ModelWrappers._get_constraint — Method

_get_constraint(param)

Recursively collect constraints of 'Param' struct. Not exported.

Examples

ModelWrappers._get_val — Method

_get_val(param)

Recursively collect values of 'Param' struct. Not exported.

Examples

ModelWrappers._paramnames — Method

_paramnames(sym, len)

Return parameter names as a string in increasing order. Not exported.

Examples

ModelWrappers._to_bijector — Method

_to_bijector(infoᵥ)

Transform user constraint to Bijector. Not exported.

Examples

ModelWrappers._to_inv_bijector — Method

_to_inv_bijector(info)

Transform Bijector to its inverse. Not exported.

Examples

ModelWrappers.check_gradients — Function

check_gradients(_rng, objective)
check_gradients(_rng, objective, ADlibraries)
check_gradients(_rng, objective, ADlibraries, θᵤ)
check_gradients(_rng, objective, ADlibraries, θᵤ, difftune; printoutput)

Check gradient computations of different backends against 'objective'.

Examples

ModelWrappers.constrain — Method

Inverse-Transform 'θᵤ' into a constrained space given 'b⁻¹'. Returns same type as 'θᵤ'.

Examples

ModelWrappers.dynamics — Method

dynamics(objective)

Assign model dynamics for a given objective.

Examples

ModelWrappers.fill_array! — Method

fill_array!(buffer, vec)

Fill array with elements of vec. Not exported.

Examples

ModelWrappers.flatten — Function

flatten(x )

Convert 'x' into a Vector.

Examples

ModelWrappers.flatten — Method

flatten(model)

Flatten 'model' values and return as vector.

Examples

ModelWrappers.flatten_Simplex — Method

Flatten vector x to k-1 dimensions. Not exported.

Examples

ModelWrappers.flatten_Symmetric — Method

flatten_Symmetric(mat, idx)

Flatten matrix to vector. Not exported.

Examples

ModelWrappers.flatten_array — Method

flatten_array(mat)

Flatten array x. Not exported.

Examples

ModelWrappers.generate — Method

generate(_rng, objective)

Generate statistics given model parameter and data.

Examples

ModelWrappers.log_abs_det_jac — Method

log_abs_det_jac(model)

Evaluate eventual Jacobian adjustments from transformations at 'model' values.

Examples

ModelWrappers.log_abs_det_jac — Method

log_abs_det_jac(b, θ)

Evaluate eventual Jacobian adjustments from transformation of 'b' at 'θ'.

Examples

ModelWrappers.log_density — Function

log_density(objective, θᵤ)

Compute log density of 'objective' at 'θᵤ'.

Examples

ModelWrappers.log_density_and_gradient — Function

log_density_and_gradient(objective, θᵤ)

Compute log density and gradient of 'objective' at 'θᵤ'.

Examples

ModelWrappers.log_density_and_hessian — Function

log_density_and_hessian(objective, θᵤ)

Compute log density, gradient and hessian of 'objective' at 'θᵤ'.

Examples

ModelWrappers.log_prior — Method

log_prior(model)

Evaluate Log density of 'model' prior given current 'model' values.

Examples

ModelWrappers.log_prior — Method

Evaluate Log density of 'prior' at 'θ'.

Examples

ModelWrappers.log_prior_with_transform — Method

log_prior_with_transform(model)

Evaluate Log density and eventual Jacobian adjustments of 'model' prior given current 'model' values.

Examples

ModelWrappers.log_prior_with_transform — Method

log_prior_with_transform(prior, θ)

Evaluate Log density and eventual Jacobian adjustments from transformation of 'prior' at 'θ'.

Examples

ModelWrappers.paramcount — Method

paramcount(sym, types, val)

Count length of nested parameter tuple. Not exported.

Examples

ModelWrappers.paramnames — Method

paramnames(sym, types, val, constraint)

Return all parameter names in increasing order. Not exported.

Examples

ModelWrappers.predict — Method

predict(_rng, objective)

Predict new data given model parameter and data.

Examples

ModelWrappers.sample_constraint — Method

sample_constraint(_rng, prior)

Sample from constraint if 'prior'.

Examples

ModelWrappers.simulate — Method

simulate(model)

Simulate data given Model parameter.

Examples

ModelWrappers.tag — Method

tag(mat)
tag(mat, upper)
tag(mat, upper, diag)

Assign subset of elements to track in Matrix mat. Not exported.

Examples

ModelWrappers.unconstrain — Method

unconstrain(model)

Unconstrain 'model' values and return as NamedTuple.

Examples

ModelWrappers.unconstrain — Method

unconstrain(b, θ)

Transform 'θ' into an unconstrained space given 'b'. Returns same type as 'θ'.

Examples

ModelWrappers.unconstrain_flatten — Method

unconstrain_flatten(model)

Flatten and unconstrain 'model' values and return as vector.

Examples

ModelWrappers.unflatten! — Method

Inplace version of unflatten.

Examples

ModelWrappers.unflatten — Method

Unlatten Vector 'θ' given constraints from 'model' and return as NamedTuple.

Examples

ModelWrappers.unflatten_constrain! — Method

Inplace version of unflatten_constrain.

Examples

ModelWrappers.unflatten_constrain — Method

unflatten_constrain(model, θᵤ)

Constrain and Unflatten vector 'θᵤ' given 'model' constraints.

Examples

StatsBase.sample! — Method

sample!(_rng, model)

Inplace version of sample.

Examples

StatsBase.sample — Method

sample(_rng, model)

Sample from 'model' prior and return as NamedTuple.

Examples